Handle Errors following clean architecture and implement Formik

+++ Application/Activities/Delete.cs
@@ -1,15 +1,16 @@
 using MediatR;
 using Domain;
 using Persistence;
+using Application.Core;
 
 namespace Application.Activities
 {
     public class Delete
     {
-        public class Command : IRequest{
+        public class Command : IRequest<Result<Unit>>{
             public Guid Id { get; set; }
         }
-        public class Handler : IRequestHandler<Command>
+        public class Handler : IRequestHandler<Command,Result<Unit>>
         {
             private readonly DataContext _context;
 
@@ -18,12 +19,14 @@ namespace Application.Activities
                 _context = context;
             }
 
-            public async Task<Unit> Handle(Command request, CancellationToken cancellationToken)
+            public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
             {
                 var activity = await _context.Activities.FindAsync(request.Id);
+                if(activity==null) return null;
                 _context.Activities.Remove(activity);
-                await _context.SaveChangesAsync();
-                return Unit.Value;
+                var result = await _context.SaveChangesAsync()>0;
+                if(!result) return Result<Unit>.Failure("Failed to delete");
+                return Result<Unit>.Success( Unit.Value);
 
             }
         }
